requirements.txt
appnope==0.1.0
asgiref==3.2.3
asn1crypto==1.3.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.3
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.5
django-nine==0.2.3
djangorestframework==3.11.0
elasticsearch==7.0.4
elasticsearch-dsl==7.1.0
entrypoints==0.3
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.3
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.5.2
pyOpenSSL==19.1.0
pyrsistent==0.15.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.1
requests==2.23.0
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
zipp==2.2.0
